feat: 🎸 add Mobile Swap
