+++ src/i18n/languages/zh_TW.json
@@ -20,7 +20,7 @@
     "balance": "餘額",
     "enterTipOfSubscription": "Please Enter the Amount of Your Subscription",
     "amountToClaim": "Amount to Claim",
-    "connectWallet": "Connect Wallet",
+    "connectWallet": "連接錢包",
     "connectWallet1": "Connect",
 
     "exchangeCurrencyTitle": "Swap AWAKEN with SASHIMI",
@@ -609,6 +609,20 @@
     "Original": "基本版",
     "Depth": "深度",
     "ensuring that you approve an adequate amount": "{{symbol}} 額度不足。請重新操作並確保您的授權金額滿足需求（{{amount}} {{symbol}}）。",
-    "My": "我的"
+    "My": "我的",
+
+    "Swap": "兌換",
+    "Pay": "支付",
+    "Receive": "得到",
+    "Insufficient liquidity for this trade": "此交易流動性不足",
+
+    "Setting a high slippage will help the transaction, but the transaction price may not be ideal.": "設定一個高滑點將有助於成交，但成交價格可能不會太理想。",
+    "Min. Received": "最小得到",
+    "Min.Received refers to the exchange result at the price corresponding to the Max.Slippage you set.Generally, it will be more.": "最小得到數量是指以你設定的最大滑點對應的價格成交的兌換結果，一般情況下會更多。",
+    "The maximum impact on the currency price of the liquidity pool after the transaction is completed.": "完成交易後對流動性池的幣價產生的最大影響。",
+    "DEX fees belong to liquidity providers and are already included in the current quote.": "DEX手續費歸屬流動性提供者，並已計入當前報價。",
+    "Txn Fee are the miner fees paid in order for transactions to proceed": "交易費是為了進行交易而支付的礦工費。",
+    "Order Routing": "訂單路由",
+    "Awaken's order routing selects the swap path with the lowest comprehensive cost to complete the transaction and increase the amount you receive.": "Awaken的訂單路由選擇綜合費用最低的兌換路徑來完成交易，增加你的得到金額。"
   }
 }
