+++ src/pages/Swap/components/SwapCircleProcess/styles.less
@@ -7,7 +7,7 @@
     width: 100%;
     height: 100%;
     border-radius: 50%;
-    border: 2px solid #9ba0b0;
+    border: 2px solid #5685ff;
     border-top-color: transparent;
     border-right-color: transparent;
     transform: rotate(45deg);
@@ -18,7 +18,7 @@
     width: 100%;
     height: 100%;
     border-radius: 50%;
-    border: 2px solid #9ba0b0;
+    border: 2px solid #5685ff;
     border-bottom-color: transparent;
     border-left-color: transparent;
     transform: rotate(45deg);
@@ -68,7 +68,7 @@
 @keyframes swap-change-color {
   0%,
   50% {
-    border-color: #9ba0b0;
+    border-color: #5685ff;
     border-bottom-color: transparent;
     border-left-color: transparent;
   }
