+++ src/pages/Swap/components/SwapInputRow/index.tsx
@@ -3,14 +3,12 @@ import BigNumber from 'bignumber.js';
 import { Col, InputProps, InputRef, Row } from 'antd';
 
 import { Currency } from '@awaken/sdk-core';
-import { CurrencyLogo } from 'components/CurrencyLogo';
 import { parseInputChange, unitConverter } from 'utils';
 import { useTranslation } from 'react-i18next';
 import { useTokenPrice } from 'contexts/useTokenPrice/hooks';
 import { isValidNumber } from 'utils/reg';
 import { divDecimals } from 'utils/calculate';
 
-import { Pair } from 'components/Pair';
 import CommonInput from 'components/CommonInput';
 import Font from 'components/Font';
 import { ZERO } from 'constants/misc';
@@ -18,6 +16,7 @@ import { ZERO } from 'constants/misc';
 import './styles.less';
 import PriceUSDDigits from 'components/PriceUSDDigits';
 import getFontStyle from 'utils/getFontStyle';
+import { useMobile } from 'utils/isMobile';
 
 interface Props extends Omit<InputProps, 'onChange'> {
   token?: Currency;
@@ -31,6 +30,8 @@ interface Props extends Omit<InputProps, 'onChange'> {
   value?: string;
   showMax?: boolean;
   gasFee?: string | number;
+  title?: string;
+  usdSuffix?: React.ReactNode;
 }
 export default function SwapInputRow(props: Props) {
   const {
@@ -44,10 +45,12 @@ export default function SwapInputRow(props: Props) {
     disabled = false,
     balance,
     showMax = false,
-    maxCallback = () => null,
     gasFee,
+    title = '',
+    usdSuffix,
   } = props;
   const { t } = useTranslation();
+  const isMobile = useMobile();
 
   const inputRef = useRef<InputRef>(null);
 
@@ -82,15 +85,22 @@ export default function SwapInputRow(props: Props) {
   const renderUsd = useMemo(() => {
     if (value === undefined || value === '')
       return (
-        <Font size={14} color="two">
+        <Font size={isMobile ? 12 : 14} color="two">
           -
         </Font>
       );
 
     return (
-      <PriceUSDDigits className={getFontStyle({ size: 14, color: 'two' })} price={ZERO.plus(value).times(tokenPrice)} />
+      <>
+        <PriceUSDDigits
+          wrapperClassName="swap-input-price-wrap"
+          className={getFontStyle({ size: isMobile ? 12 : 14, color: 'two' })}
+          price={ZERO.plus(value).times(tokenPrice)}
+        />
+        {usdSuffix}
+      </>
     );
-  }, [value, tokenPrice]);
+  }, [value, isMobile, tokenPrice, usdSuffix]);
 
   const onMax = useCallback(() => {
     if (token?.symbol === 'ELF' && gasFee && balance) {
@@ -101,33 +111,42 @@ export default function SwapInputRow(props: Props) {
   }, [balance, gasFee, setValue, token?.decimals, token?.symbol]);
 
   return (
-    <Row gutter={[0, 12]} justify="space-between" className="swap-input-row" onClick={() => inputRef.current?.focus()}>
-      <Col span={24}>
-        <CommonInput
-          suffix={suffix}
-          onChange={onInputChange}
-          value={value ?? ''}
-          placeholder={placeholder}
-          className="swap-input"
-          disabled={!token || disabled}
-          ref={inputRef}
-        />
-      </Col>
-      <Col>{!hideUSD && renderUsd}</Col>
-      <Col>
-        {!hidBlance && (
-          <div className="blance-box">
-            <Font size={14} color="two" lineHeight={20}>
-              {`${t('balance')}：${displayBalance}`}
-            </Font>
-            {showMax && (
-              <div className="max-btn" onClick={onMax}>
-                MAX
-              </div>
-            )}
-          </div>
-        )}
-      </Col>
-    </Row>
+    <div className="swap-input-row" onClick={() => inputRef.current?.focus()}>
+      <Font color="two" lineHeight={22} size={14}>
+        {title}
+      </Font>
+
+      <Row gutter={[0, 12]} justify="space-between">
+        <Col span={24}>
+          <CommonInput
+            suffix={suffix}
+            onChange={onInputChange}
+            value={value ?? ''}
+            placeholder={placeholder}
+            className="swap-input"
+            disabled={!token || disabled}
+            ref={inputRef}
+          />
+        </Col>
+      </Row>
+
+      <div className="swap-input-row-footer">
+        <div>{!hideUSD && renderUsd}</div>
+        <div>
+          {!hidBlance && (
+            <div className="balance-box">
+              <Font size={isMobile ? 12 : 14} color="two" lineHeight={20}>
+                {`${t('balance')}：${displayBalance}`}
+              </Font>
+              {showMax && (
+                <div className="max-btn" onClick={onMax}>
+                  MAX
+                </div>
+              )}
+            </div>
+          )}
+        </div>
+      </div>
+    </div>
   );
 }
