+++ src/pages/Swap/components/SwapOrderRouting/styles.less
@@ -1,8 +1,11 @@
 .swap-order-routing {
   width: 100%;
 
+  .swap-order-header {
+    margin-bottom: 18px;
+  }
+
   .swap-order-content {
-    margin-top: 18px;
     display: flex;
     flex-direction: row;
     justify-content: space-between;
