+++ src/utils/swapContract.tsx
@@ -279,8 +279,9 @@ export const onSwap: (param: SwapProps) => Promise<boolean | any> = async ({
     };
 
   const contract = routerContract;
+  const isSwap = !!path;
 
-  if (path) {
+  if (isSwap) {
     methodName = 'swapExactTokensForTokens';
     args = [bigNumberToWeb3Input(amountIn), bigNumberToWeb3Input(amountOutMin), path, account, getDeadline(), getCID()];
   } else if (tokenA?.isNative) {
@@ -324,7 +325,7 @@ export const onSwap: (param: SwapProps) => Promise<boolean | any> = async ({
       if (isUserDenied(result.error.message)) return REQ_CODE.UserDenied;
       return REQ_CODE.Fail;
     }
-    swapSuccess({ tokenB, tokenA, result, t });
+    swapSuccess({ tokenB, tokenA, result, t, isSwap });
     return REQ_CODE.Success;
   } catch (error: any) {
     formatSwapError(error, {
